created one redux store slice for all playlist action, when creating a new playlist it now show up in the list right away instead of first refreshing the page
