+++ src/App.js
@@ -5,7 +5,6 @@ import { Switch, Route } from "react-router-dom";
 import { useDispatch } from "react-redux";
 import { getUserWithStoredToken } from "./store/user/actions";
 
-import Loading from "./components/Loading";
 import NavBar from "./components/NavBar";
 import HomePage from "./pages/HomePage";
 import Login from "./pages/Login";
@@ -29,8 +28,8 @@ function App() {
   return (
     <div style={{ maxHeight: "99vh", minHeight: "99vh", overflow: "hidden" }}>
       <NavBar />
+
       <Switch>
-        <Route path="/loading" component={Loading} />
         <Route path="/signup" component={Signup} />
         <Route path="/login" component={Login} />
         <Route path="/spotifyredirect" component={SpotifyRedirect} />
