--- src/store/createPlaylist/selectors.js
@@ -1,1 +1,0 @@
-export const selectNewPlaylist = (state) => state.newPlaylist;
