+++ src/store/rootReducer.js
@@ -2,8 +2,7 @@ import { combineReducers } from "redux";
 import appState from "./appState/reducer";
 import user from "./user/reducer";
 import spotifyToken from "./spotifyToken/reducer";
-import newPlaylist from "./createPlaylist/reducer";
-import playlists from "./getPlaylists/reducer";
+import playlists from "./Playlists/reducer";
 import song from "./playSong/reducer";
 import search from "./searchInput/reducer";
 import spotifyMusic from "./spotifyMusic/reducer";
@@ -12,7 +11,6 @@ export default combineReducers({
   appState,
   user,
   spotifyToken,
-  newPlaylist,
   playlists,
   song,
   search,
